remove main screen from app module
