+++ app/src/main/java/com/giraffe/cineverseapp/MainActivity.kt
@@ -10,6 +10,7 @@ import androidx.compose.ui.graphics.toArgb
 import com.giraffe.designsystem.theme.CineVerseTheme
 import com.giraffe.explore.ExploreApi
 import com.giraffe.home.HomeApi
+import com.giraffe.home.navigation.MainScreen
 import dagger.hilt.android.AndroidEntryPoint
 import jakarta.inject.Inject
