+++ presentation/home/src/main/java/com/giraffe/home/navigation/DiscoverRoute.kt
@@ -1,16 +1,6 @@
 package com.giraffe.home.navigation
 
-//@Serializable
-// object DiscoverRoute
-//
-// fun NavController.navigateToDiscover() {
-//    navigate(DiscoverRoute)
-//}
-//
-// fun NavGraphBuilder.discoverRoute(
-//    exploreApi: ExploreApi,
-//) {
-//    composable<DiscoverRoute> {
-//        exploreApi.ExploreContainer()
-//    }
-//}
+import kotlinx.serialization.Serializable
+
+@Serializable
+object DiscoverRoute
