+++ presentation/home/src/main/java/com/giraffe/home/navigation/HomeNavGraph.kt
@@ -1,17 +1,26 @@
 package com.giraffe.home.navigation
 
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
+import com.giraffe.designsystem.composable.BottomNavigationBar
+import com.giraffe.designsystem.composable.BottomTab
 import com.giraffe.details.DetailsApi
+import com.giraffe.explore.ExploreApi
+import com.giraffe.home.HomeApi
 import com.giraffe.home.screen.home.HomeRoute
 import com.giraffe.home.screen.home.homeRoute
 import com.giraffe.home.screen.movies_list.moviesListRoute
@@ -74,3 +83,68 @@ fun HomeNavGraph(
     }
 }
 
+@Composable
+fun MainScreen(
+    exploreApi: ExploreApi,
+    homeApi: HomeApi
+) {
+    val navController: NavHostController = rememberNavController()
+
+    var isBottomBarVisible by remember { mutableStateOf(true) }
+
+    var selectedTabBottomBar by remember { mutableStateOf(BottomTab.Home) }
+
+    Column {
+        NavHost(
+            navController = navController,
+            startDestination = HomeRoute,
+            modifier = Modifier.weight(1f)
+        ) {
+
+            composable<HomeRoute> {
+                homeApi.HomeContainer {
+                    isBottomBarVisible = it
+                }
+            }
+
+            composable<DiscoverRoute> {
+                exploreApi.ExploreContainer {
+                    isBottomBarVisible = it
+                }
+            }
+        }
+
+        BottomNavigationBar(
+            selectedTab = selectedTabBottomBar,
+            isBottomBarVisible = isBottomBarVisible,
+            onTabSelected = {
+                when (it) {
+                    BottomTab.Explore -> {
+                        selectedTabBottomBar = BottomTab.Explore
+                        navController.navigate(DiscoverRoute) {
+                            popUpTo(navController.graph.startDestinationRoute ?: "") {
+                                saveState = true
+                            }
+                            launchSingleTop = true
+                            restoreState = true
+                        }
+                    }
+
+                    BottomTab.Home -> {
+                        selectedTabBottomBar = BottomTab.Home
+                        navController.navigate(HomeRoute) {
+                            popUpTo(navController.graph.startDestinationRoute ?: "") {
+                                saveState = true
+                            }
+                            launchSingleTop = true
+                            restoreState = true
+                        }
+                    }
+
+                    else -> {
+                    }
+                }
+            }
+        )
+    }
+}
